+++ src/main/java/com/arcanc/pulselib/content/registration/Registration.java
@@ -12,12 +12,15 @@ package com.arcanc.pulselib.content.registration;
 import com.arcanc.pulselib.content.registration.block.TestBlock;
 import com.arcanc.pulselib.content.registration.block.block_entity.TestBlockEntity;
 import com.arcanc.pulselib.content.registration.entity.TestEntity;
+import com.arcanc.pulselib.content.registration.item.TestArmor;
 import com.arcanc.pulselib.content.registration.item.TestBlockItem;
 import com.arcanc.pulselib.util.PLibDatabase;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.ArmorMaterials;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockBehaviour;
@@ -67,6 +70,21 @@ public class Registration
 		public static final DeferredItem<TestBlockItem> TEST_ITEM = ITEMS.register("test_block", identifier -> new TestBlockItem(
 				BlockReg.TEST_BLOCK.get(),
 				new Item.Properties()));
+		
+		public static final DeferredItem<TestArmor> TEST_CHESTPLATE = ITEMS.register("test_chest", identifier -> new TestArmor(
+				ArmorMaterials.DIAMOND,
+				ArmorItem.Type.CHESTPLATE,
+				new Item.Properties().durability(ArmorItem.Type.CHESTPLATE.getDurability(33))));
+		
+		public static final DeferredItem<TestArmor> TEST_HAT = ITEMS.register("test_hat", identifier -> new TestArmor(
+				ArmorMaterials.DIAMOND,
+				ArmorItem.Type.HELMET,
+				new Item.Properties().durability(ArmorItem.Type.HELMET.getDurability(33))));
+		
+		public static final DeferredItem<TestArmor> TEST_LEGGINGS = ITEMS.register("test_leggings", identifier -> new TestArmor(
+				ArmorMaterials.DIAMOND,
+				ArmorItem.Type.LEGGINGS,
+				new Item.Properties().durability(ArmorItem.Type.LEGGINGS.getDurability(33))));
 		
 		private static void init (@NotNull final IEventBus bus)
 		{