+++ src/main/java/com/arcanc/pulselib/content/registration/item/TestArmor.java
@@ -1,0 +1,69 @@
+/**
+ * @author ArcAnc
+ * Created at: 02.07.2026
+ * Copyright (c) 2026
+ * <p>
+ * This code is licensed under "Arc's License of Common Sense"
+ * Details can be found in the license file in the root folder of this project
+ */
+
+package com.arcanc.pulselib.content.registration.item;
+
+
+import com.arcanc.pulselib.content.renderer.modelData.PModelData;
+import com.arcanc.pulselib.util.PLibDatabase;
+import com.arcanc.pulselib.util.PRenderTypes;
+import com.arcanc.pulselib.util.armor.PulseArmorAttachment;
+import com.arcanc.pulselib.util.armor.PulseArmorDefinition;
+import com.arcanc.pulselib.util.armor.PulseArmorModels;
+import com.arcanc.pulselib.util.armor.VanillaHumanoidPart;
+import net.minecraft.core.Holder;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.ArmorMaterial;
+
+import java.util.List;
+import java.util.Set;
+
+public class TestArmor extends ArmorItem
+{
+	public static final ResourceLocation TEXTURE = PLibDatabase.rl("entity/armor/test_armor/0");
+	
+	public TestArmor(Holder<ArmorMaterial> material, Type type, Properties properties)
+	{
+		super(material, type, properties);
+		PulseArmorModels.register(this, createArmorDefinition(type));
+	}
+	
+	private static PulseArmorDefinition createArmorDefinition(Type type)
+	{
+		return new PulseArmorDefinition(
+				new PModelData.Builder(PLibDatabase.rl("armor/test_armor"), "entity").build(),
+				Set.of(type.getSlot()),
+				attachments(type),
+				true);
+	}
+	
+	private static List<PulseArmorAttachment> attachments(Type type)
+	{
+		return switch (type)
+		{
+			case HELMET -> List.of(new PulseArmorAttachment(
+					type.getSlot(),
+					VanillaHumanoidPart.HEAD,
+					"head",
+					PRenderTypes.RenderTypeProvider :: trianglesSolid));
+			case CHESTPLATE -> List.of(new PulseArmorAttachment(
+					type.getSlot(),
+					VanillaHumanoidPart.RIGHT_ARM,
+					"right_arm",
+					PRenderTypes.RenderTypeProvider :: trianglesSolid));
+			case LEGGINGS -> List.of(new PulseArmorAttachment(
+					type.getSlot(),
+					VanillaHumanoidPart.RIGHT_LEG,
+					"right_leg",
+					PRenderTypes.RenderTypeProvider :: trianglesSolid));
+			case BODY, BOOTS -> List.of();
+		};
+	}
+}