+++ changelog.md
@@ -1,1 +1,2 @@
 * added emissive texture support
+* added possibility to render custom armor on vanilla models